server_python/requirements.txt
aiosqlite==0.20.0
boto3==1.34.153
Flask==3.0.3
langchain_community==0.2.11
langchain_core==0.2.28
langchain_openai==0.1.20
langgraph==0.1.19
neo4j==5.23.0
numpy
openai==1.38.0
pymongo==4.6.3
python-dotenv==1.0.1
Requests==2.32.3
typing_extensions==4.12.2